requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
tqdm
python-dotenv
google-genai
joblib
matplotlib
seaborn
notebook
